apply clippy hints

+++ src/ui/widget/popup/exercise_popup.rs
@@ -104,10 +104,7 @@ impl Popup for ExercisePopup {
                         .daily_quest_progress
                         .insert(self.title.to_lowercase(), done);
 
-                    app.add_xp(
-                        self.exercise.category.clone(),
-                        self.exercise.xp(amount) as i32,
-                    );
+                    app.add_xp(self.exercise.category.clone(), self.exercise.xp(amount));
                     return true;
                 }
             }
